webapp/wb_app/requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
backports.zoneinfo==0.2.1
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.2
djangorestframework==3.14.0
et-xmlfile==1.1.0
frozenlist==1.3.1
idna==3.4
multidict==6.0.2
openpyxl==3.0.10
pydantic==1.10.2
python-dotenv==0.21.0
pytz==2022.4
requests==2.28.1
sqlparse==0.4.3
typing_extensions==4.4.0
urllib3==1.26.12
yarl==1.8.1
psycopg2-binary==2.9.3